+++ mio/core/parser.py
@@ -1,7 +1,8 @@
 from mio import runtime
-from mio.utils import method, Null
+from mio.parser import parse
 from mio.object import Object
-from mio.parser import parse, tokenize
+from mio.lexer import tokenize
+from mio.utils import method, Null
 
 
 class Parser(Object):